Add task state transition tests and redesign mobile UI
- Add comprehensive TaskStateTransitionTests with 54 Playwright tests
  covering all task state transitions (Waiting↔Tasks↔Scheduled)
- Fix MoveToToday bug where tasks disappeared after moving from Waiting
- Redesign mobile Planner with horizontal day-roller and section tabs
- Add TEST_INVENTORY.md documenting all 97 UI tests by feature
- Fix float conversion and add missing API methods for mobile

+++ tests/PFD.PlaywrightTests/PFD.PlaywrightTests/PlannerTests.cs
@@ -1207,9 +1207,9 @@ public class PlannerTests : PageTest
         Assert.That(box, Is.Not.Null);
 
         // Simulate horizontal scroll (swipe left)
-        var startX = box!.X + box.Width * 0.8;
-        var endX = box.X + box.Width * 0.2;
-        var centerY = box.Y + box.Height / 2;
+        var startX = (float)(box!.X + box.Width * 0.8);
+        var endX = (float)(box.X + box.Width * 0.2);
+        var centerY = (float)(box.Y + box.Height / 2);
 
         Console.WriteLine("Performing horizontal swipe...");
         await Page.Mouse.MoveAsync(startX, centerY);
